requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
PyMySQL==1.1.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
cryptography==41.0.4
email-validator==2.0.0
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
flower==2.0.1
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
marshmallow==3.20.1
flask-restful==0.3.10 